Revert "fix(server/previews): fix factory method which require getStreams as …"
This reverts commit 92712a303527b21210e2b82f5c7de7c03f8b44ed.
