+++ packages/server/package.json
@@ -155,12 +155,10 @@
     "@types/pino-http": "^5.8.4",
     "@types/request": "^2.48.12",
     "@types/sanitize-html": "^2.6.2",
-    "@types/string-pixel-width": "^1.10.3",
     "@types/supertest": "^2.0.12",
     "@types/ua-parser-js": "^0.7.39",
     "@types/uuid": "^9.0.0",
     "@types/verror": "^1.10.6",
-    "@types/xml-escape": "^1.1.3",
     "@types/yargs": "^17.0.10",
     "@types/zxcvbn": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
